[FE] refactor(#41): 프로젝트 패키지 구조 수정
