+++ frontend/app/src/main/java/com/takseha/frontend/MainActivity.kt
@@ -3,8 +3,7 @@ package com.takseha.frontend
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.takseha.presentation.HomeActivity
-import dagger.hilt.android.HiltAndroidApp
+import com.takseha.presentation.ui.HomeActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
