+++ frontend/presentation/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
             android:name=".SecActivity"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.HomeActivity"
             android:exported="false" />
     </application>
